Migrated tests to match new structure
