--- tests/foundry/client/core/mouse.test-d.ts
@@ -1,5 +1,0 @@
-import { expectTypeOf } from "vitest";
-
-// const mouseManager = new MouseManager();
-
-expectTypeOf(MouseManager.MOUSE_WHEEL_RATE_LIMIT).toEqualTypeOf<number>();
