+++ tests/foundry/client/helpers/client-issues.test-d.ts
