+++ tests/foundry/client/helpers/client-settings.test-d.ts
