+++ tests/foundry/client/helpers/document-index.test-d.ts
