+++ tests/foundry/client/helpers/interaction/client-keybindings.test-d.ts
