+++ tests/foundry/client/helpers/interaction/clipboard-helper.test-d.ts
