+++ tests/foundry/client/helpers/interaction/gamepad-manager.test-d.ts
