+++ tests/foundry/client/helpers/interaction/keyboard-manager.test-d.ts
