+++ tests/foundry/client/helpers/interaction/mouse-manager.test-d.ts
@@ -1,0 +1,5 @@
+import { expectTypeOf } from "vitest";
+
+// const mouseManager = new MouseManager();
+
+expectTypeOf(foundry.helpers.interaction.MouseManager.MOUSE_WHEEL_RATE_LIMIT).toEqualTypeOf<number>();
