+++ tests/foundry/client/helpers/interaction/tooltip-manager.test-d.ts
