+++ tests/foundry/client/helpers/media/image-helper.test-d.ts
