+++ tests/foundry/client/helpers/media/video-helper.test-d.ts
