+++ tests/foundry/client/helpers/socket-interface.test-d.ts
