+++ tests/foundry/client/helpers/time.test-d.ts
